fs/9p: Don't open remote file with APPEND mode when writeback cache is used
When page cache is used, writebacks are done on a page granularity, and it
is expected that the underlying filesystem (such as v9fs) should respect
the write position.  However, currently v9fs will passthrough O_APPEND to
the server even on cached mode.  This causes data corruption if a sync or
fstat gets between two writes to the same file.

This patch removes the APPEND flag from the open request we send to the
server when writeback caching is involved.  I believe keeping server-side
APPEND is probably fine for uncached mode (even if two fds are opened, one
without O_APPEND and one with it, this should still be fine since they
would use separate fid for the writes).

Signed-off-by: Tingmao Wang <m@maowtm.org>
---

I haven't done a bisect to figure out if this regression was introduced by
a change or was this behaviour always present - 6.14 has the same problem
and 6.13 did not compile for me due to some problem with bool / true /
false being a keyword in c23, and it could not compile with c17 either.