selftests/landlock: Repurpose scoped_abstract_unix_test.c for pathname sockets too
Since there is very little difference between abstract and pathname
sockets in terms of testing of the scoped access checks (the only
difference is in which scope bit control which form of socket), it makes
sense to reuse the existing test for both type of sockets.  Therefore,
we rename scoped_abstract_unix_test.c to scoped_unix_test.c and extend
the scoped_domains test to test pathname (i.e. non-abstract) sockets
too.

Since we can't change the variant data of scoped_domains (as it is
defined in the shared .h file), we do this by extracting the actual test
code into a function, and call it from different test cases.

Also extend scoped_audit (this time we can use variants) to test both
abstract and pathname sockets.  For pathname sockets, audit_log_lsm_data
will produce path="..." (or hex if path contains control characters)
with absolute paths from the dentry, so we need to construct the escaped
regex for the real path like in fs_test.

Signed-off-by: Tingmao Wang <m@maowtm.org>
---

Changes in v3:
- Fix comment, style, formatting, and reflow the above message
- Update definition of log_match_remaining